Allow a git clone to run in place.
These are the python modules that are needed to run the simpleblend program.
With these in the repository you should be able to just run simpleblend
(and it will find these modules for you) and not have to install any UFO3
branches of python modules yourself.

With these a local git